+++ lib/MakerEnhanceClient.d.ts
@@ -1,0 +1,13 @@
+/// <reference types="react" />
+interface MakerEnhanceClientProps {
+    user: string;
+    index?: number;
+    scriptSrc: string;
+}
+declare global {
+    interface Window {
+        MakerEmbeds: any;
+    }
+}
+export default function useMakerEnhance({ user, index, scriptSrc }: MakerEnhanceClientProps): JSX.Element;
+export {};
